refactor and format
