+++ .prettierrc
@@ -1,7 +1,7 @@
 {
     "printWidth": 80,
     "useTabs": false,
-    "tabWidth": 2,
+    "tabWidth": 4,
     "trailingComma": "es5",
     "semi": true,
     "singleQuote": false,
