+++ src/App.tsx
@@ -5,25 +5,27 @@ import AppRoutes from "./Routes";
 import "./App.scss";
 
 function App() {
-  const isDarkMode = useSelector((state: AppStore) => state.ui.isDarkMode);
-  const direction = useSelector((state: AppStore) => state.ui.direction);
+    const isDarkMode = useSelector((state: AppStore) => state.ui.isDarkMode);
+    const direction = useSelector((state: AppStore) => state.ui.direction);
 
-  return (
-    <>
-      <ConfigProvider
-        direction={direction}
-        theme={{
-          algorithm: isDarkMode ? theme.darkAlgorithm : theme.defaultAlgorithm,
-          // token: { fontFamily: "Zain" },
-        }}>
-        <div
-          className={isDarkMode ? "App App--dark" : "App App--light"}
-          dir={direction}>
-          <AppRoutes />
-        </div>
-      </ConfigProvider>
-    </>
-  );
+    return (
+        <>
+            <ConfigProvider
+                direction={direction}
+                theme={{
+                    algorithm: isDarkMode
+                        ? theme.darkAlgorithm
+                        : theme.defaultAlgorithm,
+                    // token: { fontFamily: "Zain" },
+                }}>
+                <div
+                    className={isDarkMode ? "App App--dark" : "App App--light"}
+                    dir={direction}>
+                    <AppRoutes />
+                </div>
+            </ConfigProvider>
+        </>
+    );
 }
 
 export default App;
