+++ src/components/ClassDetailsCard/ClassDetailsCard.scss
@@ -1,0 +1,4 @@
+// outline to help with debugging, remove later
+.out-line {
+    outline: 1px solid yellow;
+}
