+++ src/components/ExportModal/DownloadModal.scss
@@ -1,5 +1,4 @@
 .modal {
-
     &__content {
         padding: 1.5rem;
         height: 10.75rem;
