+++ src/components/ExportModal/DownloadModal.tsx
@@ -8,62 +8,62 @@ import "./DownloadModal.scss";
 const { Text } = Typography;
 
 interface DownloadModalProps {
-  title: string;
-  dataSelectorFunction: any; //NOT SURE WHAT THE TYPE SHOULD BE
+    title: string;
+    dataSelectorFunction: any; //NOT SURE WHAT THE TYPE SHOULD BE
 }
 
 const DownloadModal = ({ title, dataSelectorFunction }: DownloadModalProps) => {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const { t } = useTranslation();
+    const [isModalOpen, setIsModalOpen] = useState(false);
+    const { t } = useTranslation();
 
-  const showModal = () => {
-    setIsModalOpen(true);
-  };
+    const showModal = () => {
+        setIsModalOpen(true);
+    };
 
-  const handleOk = () => {
-    //DOWNLOAD LOGIC GOES HERE
-    console.log("Downloading...");
-    setIsModalOpen(false);
-  };
+    const handleOk = () => {
+        //DOWNLOAD LOGIC GOES HERE
+        console.log("Downloading...");
+        setIsModalOpen(false);
+    };
 
-  const handleCancel = () => {
-    setIsModalOpen(false);
-  };
-  return (
-    <>
-      <Button onClick={showModal} icon={<DownloadOutlined />}></Button>
-      <Modal
-        className="modal"
-        title={t("modal.title") + title}
-        closable={{ "aria-label": "Custom Close Button" }}
-        open={isModalOpen}
-        onOk={handleOk}
-        onCancel={handleCancel}
-        cancelText={t("modal.cancel")}
-        okText={t("modal.download")}
-        okButtonProps={{ icon: <DownloadOutlined /> }}>
-        <div className="modal__content">
-          <div className="modal__content__time-frame-select">
-            <Segmented
-              options={[
-                t("modal.day"),
-                t("modal.week"),
-                t("modal.month"),
-                t("modal.all"),
-              ]}
-            />
-          </div>
-          <div className="modal__content__file-format-text">
-            <Text strong>{t("modal.selectFormat")}</Text>
-            <Text type="secondary"> • {t("modal.bodyText")}</Text>
-          </div>
-          <div className="modal__content__file-format-select">
-            <Segmented options={["PDF", "CSV"]} />
-          </div>
-        </div>
-      </Modal>
-    </>
-  );
+    const handleCancel = () => {
+        setIsModalOpen(false);
+    };
+    return (
+        <>
+            <Button onClick={showModal} icon={<DownloadOutlined />}></Button>
+            <Modal
+                className="modal"
+                title={t("modal.title") + title}
+                closable={{ "aria-label": "Custom Close Button" }}
+                open={isModalOpen}
+                onOk={handleOk}
+                onCancel={handleCancel}
+                cancelText={t("modal.cancel")}
+                okText={t("modal.download")}
+                okButtonProps={{ icon: <DownloadOutlined /> }}>
+                <div className="modal__content">
+                    <div className="modal__content__time-frame-select">
+                        <Segmented
+                            options={[
+                                t("modal.day"),
+                                t("modal.week"),
+                                t("modal.month"),
+                                t("modal.all"),
+                            ]}
+                        />
+                    </div>
+                    <div className="modal__content__file-format-text">
+                        <Text strong>{t("modal.selectFormat")}</Text>
+                        <Text type="secondary"> • {t("modal.bodyText")}</Text>
+                    </div>
+                    <div className="modal__content__file-format-select">
+                        <Segmented options={["PDF", "CSV"]} />
+                    </div>
+                </div>
+            </Modal>
+        </>
+    );
 };
 
 export default DownloadModal;
