+++ src/components/Footer/Footer.scss
@@ -1,5 +1,5 @@
 .footer {
-  padding: 20px;
-  text-align: center;
-  border-top: 1px solid #e9ecef;
+    padding: 20px;
+    text-align: center;
+    border-top: 1px solid #e9ecef;
 }
