+++ src/components/Footer/Footer.tsx
@@ -2,18 +2,18 @@ import React from "react";
 // footer component goes here
 import "./Footer.scss";
 const Footer: React.FC = () => {
-  return (
-    <div className="footer">
-      Footer component<hr></hr>
-      <hr />
-      <hr />
-      <hr />
-      <hr />
-      <hr />
-      <hr />
-      <hr />
-    </div>
-  );
+    return (
+        <div className="footer">
+            Footer component<hr></hr>
+            <hr />
+            <hr />
+            <hr />
+            <hr />
+            <hr />
+            <hr />
+            <hr />
+        </div>
+    );
 };
 
 export default Footer;
