+++ src/components/LanguageSelect/LanguageSelect.tsx
@@ -6,23 +6,23 @@ import { setDirection } from "../../store/ui.slice";
 const { Option } = Select;
 
 const LanguageSelect = () => {
-  const { i18n } = useTranslation();
-  const dispatch = useDispatch();
+    const { i18n } = useTranslation();
+    const dispatch = useDispatch();
 
-  const handleChange = (lang: "ar" | "en") => {
-    i18n.changeLanguage(lang); // Change language in i18next
-    dispatch(setDirection(lang === "ar" ? "rtl" : "ltr"));
-  };
+    const handleChange = (lang: "ar" | "en") => {
+        i18n.changeLanguage(lang); // Change language in i18next
+        dispatch(setDirection(lang === "ar" ? "rtl" : "ltr"));
+    };
 
-  return (
-    <Select
-      value={i18n.language as "ar" | "en"}
-      style={{ width: 120 }}
-      onChange={handleChange}>
-      <Option value="en">English</Option>
-      <Option value="ar">العربية</Option>
-    </Select>
-  );
+    return (
+        <Select
+            value={i18n.language as "ar" | "en"}
+            style={{ width: 120 }}
+            onChange={handleChange}>
+            <Option value="en">English</Option>
+            <Option value="ar">العربية</Option>
+        </Select>
+    );
 };
 
 export default LanguageSelect;
