+++ src/components/NavBar/NavBar.scss
@@ -1,78 +1,78 @@
 @import "../../_variables";
 
 .navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0px 16px;
-  box-shadow: 0 1px 0px #f0f0f0;
-  margin: 0;
-  font-family: "Roboto", sans-serif;
-  font-size: 14px;
-  font-weight: 400;
-  height: 64px;
-
-  &__left {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0px 16px;
+    box-shadow: 0 1px 0px #f0f0f0;
     margin: 0;
-    padding: 0;
-    flex: 1;
-  }
+    font-family: "Roboto", sans-serif;
+    font-size: 14px;
+    font-weight: 400;
+    height: 64px;
 
-  &__links {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-
-  &__item {
-    cursor: pointer;
-    padding: 25px 20px;
-    transition: color 0.3s ease;
-    position: relative;
-    border-bottom: 2px solid transparent;
-
-    &__title {
-      display: flex;
-      gap: 10px;
-
-      &:hover {
-        color: #1890ff;
-        border-bottom-color: #1890ff;
-      }
+    &__left {
+        margin: 0;
+        padding: 0;
+        flex: 1;
     }
-  }
 
-  &__toolbar {
-    display: flex;
-    align-items: center;
+    &__links {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+    }
 
     &__item {
-      display: flex;
-      padding: 0px 12px;
-      align-items: center;
+        cursor: pointer;
+        padding: 25px 20px;
+        transition: color 0.3s ease;
+        position: relative;
+        border-bottom: 2px solid transparent;
+
+        &__title {
+            display: flex;
+            gap: 10px;
+
+            &:hover {
+                color: #1890ff;
+                border-bottom-color: #1890ff;
+            }
+        }
     }
 
-    &__profilePicture {
-      width: 24px;
-      height: 24px;
-      padding-right: 8px;
-      box-sizing: content-box;
+    &__toolbar {
+        display: flex;
+        align-items: center;
+
+        &__item {
+            display: flex;
+            padding: 0px 12px;
+            align-items: center;
+        }
+
+        &__profilePicture {
+            width: 24px;
+            height: 24px;
+            padding-right: 8px;
+            box-sizing: content-box;
+        }
     }
-  }
 }
 @media only screen and (max-width: $breakpoint-md) {
-  .navbar {
-    &__left {
-      display: none;
+    .navbar {
+        &__left {
+            display: none;
+        }
     }
-  }
 }
 @media only screen and (max-width: $breakpoint-lg) {
-  .navbar {
-    &__item {
-      &__icon {
-        display: none;
-      }
+    .navbar {
+        &__item {
+            &__icon {
+                display: none;
+            }
+        }
     }
-  }
 }
