+++ src/components/NavBar/NavBar.tsx
@@ -6,18 +6,18 @@ import NavBarLinks from "./NavBarLinks";
 import "./NavBar.scss";
 
 const NavBar: React.FC = () => {
-  return (
-    <>
-      <nav className="navbar">
-        <NavBarLinks />
-        <div className="navbar__toolbar">
-          <DarkModeToggle />
-          <UserDropDown />
-          <LanguageSelect />
-        </div>
-      </nav>
-    </>
-  );
+    return (
+        <>
+            <nav className="navbar">
+                <NavBarLinks />
+                <div className="navbar__toolbar">
+                    <DarkModeToggle />
+                    <UserDropDown />
+                    <LanguageSelect />
+                </div>
+            </nav>
+        </>
+    );
 };
 
 export default NavBar;
