+++ src/components/NavBar/NavPermissions.tsx
@@ -1,11 +1,11 @@
 import React, { JSX } from "react";
 import {
-  UserOutlined,
-  TeamOutlined,
-  ApartmentOutlined,
-  ContactsOutlined,
-  BarChartOutlined,
-  SettingOutlined,
+    UserOutlined,
+    TeamOutlined,
+    ApartmentOutlined,
+    ContactsOutlined,
+    BarChartOutlined,
+    SettingOutlined,
 } from "@ant-design/icons";
 
 import { NavigateFunction } from "react-router-dom";
@@ -14,144 +14,144 @@ import { TFunction } from "i18next";
 import { Paths } from "../../Routes";
 
 type NavPermissionsType = {
-  [key in UserRole | "non"]?: {
-    name: string;
-    icon: React.ElementType;
-    dropdown?: { key: string; label: JSX.Element }[];
-  }[];
+    [key in UserRole | "non"]?: {
+        name: string;
+        icon: React.ElementType;
+        dropdown?: { key: string; label: JSX.Element }[];
+    }[];
 };
 
 export const NavPermissions = (
-  navigate: NavigateFunction,
-  t: TFunction<"translation", undefined>
+    navigate: NavigateFunction,
+    t: TFunction<"translation", undefined>
 ) => {
-  const permissions: NavPermissionsType = {
-    admin: [
-      {
-        name: "student",
-        icon: UserOutlined,
-        dropdown: [
-          {
-            key: "admin-create-student",
-            label: (
-              <a onClick={() => navigate(Paths.STUDENT.CREATE)}>
-                {t("navBar.create")}
-              </a>
-            ),
-          },
-          {
-            key: "admin-find-student",
-            label: (
-              <a onClick={() => navigate(Paths.STUDENT.FIND)}>
-                {t("navBar.find")}
-              </a>
-            ),
-          },
+    const permissions: NavPermissionsType = {
+        admin: [
+            {
+                name: "student",
+                icon: UserOutlined,
+                dropdown: [
+                    {
+                        key: "admin-create-student",
+                        label: (
+                            <a onClick={() => navigate(Paths.STUDENT.CREATE)}>
+                                {t("navBar.create")}
+                            </a>
+                        ),
+                    },
+                    {
+                        key: "admin-find-student",
+                        label: (
+                            <a onClick={() => navigate(Paths.STUDENT.FIND)}>
+                                {t("navBar.find")}
+                            </a>
+                        ),
+                    },
+                ],
+            },
+            {
+                name: "teacher",
+                icon: TeamOutlined,
+                dropdown: [
+                    {
+                        key: "admin-create-teacher",
+                        label: (
+                            <a onClick={() => navigate(Paths.TEACHER.CREATE)}>
+                                {t("navBar.create")}
+                            </a>
+                        ),
+                    },
+                    {
+                        key: "admin-find-teacher",
+                        label: (
+                            <a onClick={() => navigate(Paths.TEACHER.FIND)}>
+                                {t("navBar.find")}
+                            </a>
+                        ),
+                    },
+                ],
+            },
+            {
+                name: "class",
+                icon: ApartmentOutlined,
+                dropdown: [
+                    {
+                        key: "admin-create-class",
+                        label: (
+                            <a onClick={() => navigate(Paths.CLASS.CREATE)}>
+                                {t("navBar.create")}
+                            </a>
+                        ),
+                    },
+                    {
+                        key: "admin-find-class",
+                        label: (
+                            <a onClick={() => navigate(Paths.CLASS.FIND)}>
+                                {t("navBar.find")}
+                            </a>
+                        ),
+                    },
+                ],
+            },
+            {
+                name: "user",
+                icon: ContactsOutlined,
+                dropdown: [
+                    {
+                        key: "admin-create-user",
+                        label: (
+                            <a onClick={() => navigate(Paths.USER.CREATE)}>
+                                {t("navBar.create")}
+                            </a>
+                        ),
+                    },
+                    {
+                        key: "admin-find-user",
+                        label: (
+                            <a onClick={() => navigate(Paths.USER.FIND)}>
+                                {t("navBar.find")}
+                            </a>
+                        ),
+                    },
+                ],
+            },
+            { name: "reports", icon: BarChartOutlined },
+            { name: "system settings", icon: SettingOutlined },
         ],
-      },
-      {
-        name: "teacher",
-        icon: TeamOutlined,
-        dropdown: [
-          {
-            key: "admin-create-teacher",
-            label: (
-              <a onClick={() => navigate(Paths.TEACHER.CREATE)}>
-                {t("navBar.create")}
-              </a>
-            ),
-          },
-          {
-            key: "admin-find-teacher",
-            label: (
-              <a onClick={() => navigate(Paths.TEACHER.FIND)}>
-                {t("navBar.find")}
-              </a>
-            ),
-          },
+        teacher: [
+            {
+                name: "student",
+                icon: UserOutlined,
+                dropdown: [
+                    {
+                        key: "teacher-find-student",
+                        label: (
+                            <a onClick={() => navigate(Paths.STUDENT.FIND)}>
+                                {t("navBar.find")}
+                            </a>
+                        ),
+                    },
+                ],
+            },
+            {
+                name: "class",
+                icon: ApartmentOutlined,
+                dropdown: [
+                    {
+                        key: "teacher-find-class",
+                        label: (
+                            <a onClick={() => navigate(Paths.CLASS.FIND)}>
+                                {t("navBar.find")}
+                            </a>
+                        ),
+                    },
+                ],
+            },
+            { name: "reports", icon: BarChartOutlined },
+            { name: "settings", icon: SettingOutlined },
         ],
-      },
-      {
-        name: "class",
-        icon: ApartmentOutlined,
-        dropdown: [
-          {
-            key: "admin-create-class",
-            label: (
-              <a onClick={() => navigate(Paths.CLASS.CREATE)}>
-                {t("navBar.create")}
-              </a>
-            ),
-          },
-          {
-            key: "admin-find-class",
-            label: (
-              <a onClick={() => navigate(Paths.CLASS.FIND)}>
-                {t("navBar.find")}
-              </a>
-            ),
-          },
-        ],
-      },
-      {
-        name: "user",
-        icon: ContactsOutlined,
-        dropdown: [
-          {
-            key: "admin-create-user",
-            label: (
-              <a onClick={() => navigate(Paths.USER.CREATE)}>
-                {t("navBar.create")}
-              </a>
-            ),
-          },
-          {
-            key: "admin-find-user",
-            label: (
-              <a onClick={() => navigate(Paths.USER.FIND)}>
-                {t("navBar.find")}
-              </a>
-            ),
-          },
-        ],
-      },
-      { name: "reports", icon: BarChartOutlined },
-      { name: "system settings", icon: SettingOutlined },
-    ],
-    teacher: [
-      {
-        name: "student",
-        icon: UserOutlined,
-        dropdown: [
-          {
-            key: "teacher-find-student",
-            label: (
-              <a onClick={() => navigate(Paths.STUDENT.FIND)}>
-                {t("navBar.find")}
-              </a>
-            ),
-          },
-        ],
-      },
-      {
-        name: "class",
-        icon: ApartmentOutlined,
-        dropdown: [
-          {
-            key: "teacher-find-class",
-            label: (
-              <a onClick={() => navigate(Paths.CLASS.FIND)}>
-                {t("navBar.find")}
-              </a>
-            ),
-          },
-        ],
-      },
-      { name: "reports", icon: BarChartOutlined },
-      { name: "settings", icon: SettingOutlined },
-    ],
-    parent: [],
-    student: [],
-  };
-  return permissions;
+        parent: [],
+        student: [],
+    };
+    return permissions;
 };
