--- src/components/StudentInfo/StudentInfo.scss
@@ -1,11 +1,0 @@
-.student-section {
-  &__info {
-    margin-bottom: 1.5rem;
-  }
-  &__stat {
-    text-align: center;
-    &__progress {
-      margin-bottom: 0.5rem;
-    }
-  }
-}
