+++ src/layouts/MainLayout.scss
@@ -1,21 +1,21 @@
 @import "../_variables";
 
 .layout {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
 
-  &__content {
-    flex: 1;
-    margin: 2rem 20%;
-    min-height: 85dvh;
-    outline: 1px solid white;
-  }
+    &__content {
+        flex: 1;
+        margin: 2rem 20%;
+        min-height: 85dvh;
+        outline: 1px solid white;
+    }
 }
 @media only screen and (max-width: $breakpoint-md) {
-  .layout {
-    &__content {
-      margin: 1rem 1rem;
+    .layout {
+        &__content {
+            margin: 1rem 1rem;
+        }
     }
-  }
 }
