+++ src/locales/ar/translation.json
@@ -1,46 +1,45 @@
 {
-  "app": {
-    "title": "حلقة",
-    "description": "مدرسة القرآن الكريم"
-  },
-  "navBar": {
-    "student": "طالب",
-    "teacher": "مدرس",
-    "class": "فصل",
-    "user": "مستخدم",
-    "reports": "تقارير",
-    "system settings": "إعدادات النظام",
-    "settings": "إعدادات",
-    "logout": "الخروج",
-    "create": "إنشاء",
-    "find": "وَجَد"
-  },
-  "modal": {
-    "day": "يوم",
-    "week": "أسبوع",
-    "month": "شهر",
-    "all": "كل",
-    "download": "تحميل",
-    "cancel": "يلغي",
-    "bodyText": "سوف تحتاج إلى برنامج Excel لعرض ملف CSV",
-    "selectFormat": "حدد التنسيق",
-    "title": "تنزيل التقارير للطالب: "
-  },
-  "class": {
-    "classID": "معرف الفصل",
-    "description": "وصف",
-    "teacher": "مدرس",
-    "ageGroup": "الفئة العمرية",
-    "startsAt": "يبدأ في",
-    "endsAt": "ينتهي عند",
-    "classSize": "حجم الفصل",
-    "students": "طلاب"
-
-  },
-  "welcome": "مرحبًا",
-  "goodbye": "وداعًا",
-  "thankYou": "شكرًا لك",
-  "hello": "مرحبًا",
-  "login": "تسجيل الدخول",
-  "logout": "تسجيل الخروج"
+    "app": {
+        "title": "حلقة",
+        "description": "مدرسة القرآن الكريم"
+    },
+    "navBar": {
+        "student": "طالب",
+        "teacher": "مدرس",
+        "class": "فصل",
+        "user": "مستخدم",
+        "reports": "تقارير",
+        "system settings": "إعدادات النظام",
+        "settings": "إعدادات",
+        "logout": "الخروج",
+        "create": "إنشاء",
+        "find": "وَجَد"
+    },
+    "modal": {
+        "day": "يوم",
+        "week": "أسبوع",
+        "month": "شهر",
+        "all": "كل",
+        "download": "تحميل",
+        "cancel": "يلغي",
+        "bodyText": "سوف تحتاج إلى برنامج Excel لعرض ملف CSV",
+        "selectFormat": "حدد التنسيق",
+        "title": "تنزيل التقارير للطالب: "
+    },
+    "class": {
+        "classID": "معرف الفصل",
+        "description": "وصف",
+        "teacher": "مدرس",
+        "ageGroup": "الفئة العمرية",
+        "startsAt": "يبدأ في",
+        "endsAt": "ينتهي عند",
+        "classSize": "حجم الفصل",
+        "students": "طلاب"
+    },
+    "welcome": "مرحبًا",
+    "goodbye": "وداعًا",
+    "thankYou": "شكرًا لك",
+    "hello": "مرحبًا",
+    "login": "تسجيل الدخول",
+    "logout": "تسجيل الخروج"
 }
