+++ src/main.scss
@@ -1,3 +1,3 @@
 * {
-  font-family: "Cairo", sans-serif;
+    font-family: "Cairo", sans-serif;
 }
