+++ src/pages/error/ServerError.tsx
@@ -3,21 +3,23 @@ import "./errors.scss";
 import { useNavigate } from "react-router-dom";
 import { Paths } from "../../Routes";
 function ServerError() {
-  const navigate = useNavigate();
+    const navigate = useNavigate();
 
-  return (
-    <Result
-      className="error-page"
-      status="500"
-      title="500"
-      subTitle="Sorry, something went wrong on our end."
-      extra={
-        <Button type="primary" onClick={() => navigate(Paths.HOME.ROOT)}>
-          Back Home
-        </Button>
-      }
-    />
-  );
+    return (
+        <Result
+            className="error-page"
+            status="500"
+            title="500"
+            subTitle="Sorry, something went wrong on our end."
+            extra={
+                <Button
+                    type="primary"
+                    onClick={() => navigate(Paths.HOME.ROOT)}>
+                    Back Home
+                </Button>
+            }
+        />
+    );
 }
 
 export default ServerError;
