+++ src/pages/error/UnknownError.tsx
@@ -3,24 +3,24 @@ import "./errors.scss";
 import { useNavigate } from "react-router-dom";
 import { Paths } from "../../Routes";
 function UnknownError() {
-  const navigate = useNavigate();
+    const navigate = useNavigate();
 
-  return (
-    <Result
-      className="error-page"
-      status="error"
-      title="Something Went Wrong"
-      subTitle="An unexpected error has occurred. Please try again later."
-      extra={[
-        <Button
-          key="navigate"
-          type="primary"
-          onClick={() => navigate(Paths.HOME.ROOT)}>
-          Reload
-        </Button>,
-      ]}
-    />
-  );
+    return (
+        <Result
+            className="error-page"
+            status="error"
+            title="Something Went Wrong"
+            subTitle="An unexpected error has occurred. Please try again later."
+            extra={[
+                <Button
+                    key="navigate"
+                    type="primary"
+                    onClick={() => navigate(Paths.HOME.ROOT)}>
+                    Reload
+                </Button>,
+            ]}
+        />
+    );
 }
 
 export default UnknownError;
