+++ src/pages/error/errors.scss
@@ -1,3 +1,3 @@
 .error-page {
-  min-height: 100vh;
+    min-height: 100vh;
 }
