+++ src/pages/login/Login.scss
@@ -1,9 +1,9 @@
 .page {
-  min-height: 100vh;
+    min-height: 100vh;
 }
 .button {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
 }
