+++ src/pages/login/Login.tsx
@@ -11,26 +11,26 @@ import { Paths } from "../../Routes";
 // place login component here
 
 const Login: React.FC = () => {
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
-  const { t } = useTranslation();
+    const dispatch = useDispatch();
+    const navigate = useNavigate();
+    const { t } = useTranslation();
 
-  console.log(t("login"));
-  return (
-    <div className="page">
-      <LanguageSelect />
-      <Button
-        color="cyan"
-        variant="solid"
-        className="button"
-        onClick={() => {
-          dispatch(login());
-          navigate(Paths.HOME.ROOT);
-        }}>
-        {t("login")}
-      </Button>
-    </div>
-  );
+    console.log(t("login"));
+    return (
+        <div className="page">
+            <LanguageSelect />
+            <Button
+                color="cyan"
+                variant="solid"
+                className="button"
+                onClick={() => {
+                    dispatch(login());
+                    navigate(Paths.HOME.ROOT);
+                }}>
+                {t("login")}
+            </Button>
+        </div>
+    );
 };
 
 export default Login;
