+++ src/store/auth.slice.ts
@@ -2,26 +2,26 @@
 import { createSlice } from "@reduxjs/toolkit";
 
 interface AuthState {
-  isAuthenticated: boolean;
+    isAuthenticated: boolean;
 }
 
 const initialState: AuthState = {
-  isAuthenticated: localStorage.getItem("isAuthenticated") === "true", // Load from localStorage
+    isAuthenticated: localStorage.getItem("isAuthenticated") === "true", // Load from localStorage
 };
 
 const authSlice = createSlice({
-  name: "auth",
-  initialState,
-  reducers: {
-    login(state) {
-      state.isAuthenticated = true;
-      localStorage.setItem("isAuthenticated", "true");
+    name: "auth",
+    initialState,
+    reducers: {
+        login(state) {
+            state.isAuthenticated = true;
+            localStorage.setItem("isAuthenticated", "true");
+        },
+        logout(state) {
+            state.isAuthenticated = false;
+            localStorage.removeItem("isAuthenticated");
+        },
     },
-    logout(state) {
-      state.isAuthenticated = false;
-      localStorage.removeItem("isAuthenticated");
-    },
-  },
 });
 
 export const { login, logout } = authSlice.actions;
