+++ src/store/dummy_data/class.dummy.ts
@@ -1,6 +1,5 @@
 import { ClassState } from "../class.slice";
 
-
 export const dummyClass: ClassState = {
     id: "54372",
     description: "Main afternoon class.",
@@ -8,7 +7,5 @@ export const dummyClass: ClassState = {
     ageGroup: "6 - 10",
     startTime: "3:00 PM",
     endTime: "6:00 PM",
-    classSize: 12
-
+    classSize: 12,
 };
-
