+++ src/utils/i18n.ts
@@ -5,22 +5,21 @@ import LanguageDetector from "i18next-browser-languagedetector";
 import translationEN from "../locales/en/translation.json";
 import translationAR from "../locales/ar/translation.json";
 console.log(translationEN);
-i18n
-  .use(LanguageDetector) // detect user language
-  .use(initReactI18next) // passes i18n down to react-i18next
-  .init({
-    resources: {
-      en: {
-        translation: translationEN,
-      },
-      ar: {
-        translation: translationAR,
-      },
-    },
-    fallbackLng: "en",
-    interpolation: {
-      escapeValue: false,
-    },
-  });
+i18n.use(LanguageDetector) // detect user language
+    .use(initReactI18next) // passes i18n down to react-i18next
+    .init({
+        resources: {
+            en: {
+                translation: translationEN,
+            },
+            ar: {
+                translation: translationAR,
+            },
+        },
+        fallbackLng: "en",
+        interpolation: {
+            escapeValue: false,
+        },
+    });
 
 export default i18n;
